PR review fixes - renaming files, dom-element
